Opened Pandora's box a bit too early... I tried to rework the hit system (around HitInstance), but it ended up being a job waaaaytoo big for our current goals I don't think I'll come back to this branch, but it might be usefull as a reference later on

+++ Assets/Proto_AutoBattler/Scripts/Unit/UnitData.cs
@@ -10,7 +10,7 @@ public class UnitData : ScriptableObject
     [Min(1)] public int baseHP;
     [Min(0)] public float baseSpeed;
     [Min(0)] public float targetingRange;
-    [Min(0)] public float hitDamage;
+    [Min(0)] public int hitDamage;
     [Min(1)] public int scrapValue;
     [Min(0)] public float activationDelay;
